MCR-3497 Refactor to use subclasses per method call
